Migrate the typography components to align with Phlex conversions
